remove the test feature
