+++ rust/constants/Cargo.toml
@@ -5,15 +5,8 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
-hex = {version = "0.4.3", optional = true}
-
-[features]
-default = ["test"]
-
-active = []
-signer = []
-test = ["active", "signer", "hex"]
+[dev-dependencies]
+hex = "0.4"
 
 [lib]
 name = "constants"
