+++ rust/db_handling/Cargo.toml
@@ -28,10 +28,9 @@ pretty_assertions = { version = "1.3" }
 tempfile = { version = "3.4" }
 
 [features]
-default = ["test"]
-active = ["constants/active", "defaults/active", "definitions/active", "tiny-bip39", "sp-core", "sp-runtime", "zeroize"]
-signer = ["constants/signer", "defaults/signer", "definitions/signer", "plot_icon", "tiny-bip39", "sp-core", "sp-runtime", "zeroize"]
-test = ["active", "constants/test", "signer"]
+default = []
+active = ["defaults/active", "definitions/active", "tiny-bip39", "sp-core", "sp-runtime", "zeroize"]
+signer = ["defaults/signer", "definitions/signer", "plot_icon", "tiny-bip39", "sp-core", "sp-runtime", "zeroize"]
 
 [lib]
 name = "db_handling"
