+++ rust/definitions/Cargo.toml
@@ -22,10 +22,8 @@ sp-runtime = {git = "https://github.com/paritytech/substrate", default-features 
 sp-version = {git = "https://github.com/paritytech/substrate"}
 sp-wasm-interface = {git = "https://github.com/paritytech/substrate", optional = true}
 thiserror = "1.0.38"
-variant_count = {version = "1.1.0", optional = true}
 
 [features]
-default = ["test"]
 active = [
 	"sc-executor-wasmi",
 	"sp-io",
@@ -33,11 +31,6 @@ active = [
 ]
 signer = [
 	"plot_icon",
-]
-test = [
-	"active",
-	"signer",
-	"variant_count",
 ]
 
 [lib]
