+++ rust/parser/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2018"
 
 [dependencies]
 bitvec = "1.0.1"
-defaults = {path = "../defaults", default-features = false, optional = true}
 definitions = {path = "../definitions", default-features = false, features = ["signer"]}
 frame-metadata = { version = "15.0.0", default-features = false, features = ["std", "legacy"]}
 hex = "0.4.3"
@@ -22,14 +21,11 @@ sp-arithmetic = {git = "https://github.com/paritytech/substrate"}
 sp-core = {git = "https://github.com/paritytech/substrate", default-features = false, features = ["full_crypto"]}
 sp-runtime = {git = "https://github.com/paritytech/substrate", default-features = false}
 thiserror = "1.0.38"
-pretty_assertions = { version = "1.3.0", optional = true }
 
 [dev-dependencies]
+pretty_assertions = "1"
+defaults = { path = "../defaults", default-features = false }
 subxt = { version = "0.27", default-features = false }
-
-[features]
-default = ["test"]
-test = ["defaults/active", "pretty_assertions"]
 
 [lib]
 name = "parser"
